requirements.txt
transformers==4.27.1
googletrans==3.1.0a0
reportlab==3.6.13
img2pdf==0.4.4
PyPDF2==2.6.0 
streamlit==1.18.0
pillow==9.3.0
pdfminer.six==20221105
streamlit_drawable_canvas==0.9.2
tabula-py==2.7.0
svgpathtools==1.6.0
torch
librosa
